src/meal_planner/ml/ocr/marker_engine.py
"""Marker OCR engine for image-based OCR."""

import time
from pathlib import Path
from typing import Set

from loguru import logger
from PIL import Image

from meal_planner.core.models import OCREngine, OCRResult
from meal_planner.ml.ocr.base import BaseOCREngine


class MarkerEngine(BaseOCREngine):
    """OCR engine for image-based OCR.
    
    This is a placeholder implementation that would typically use
    a library like Tesseract, EasyOCR, or a cloud-based OCR service.
    """
    
    def __init__(self):
        """Initialize the Marker engine."""
        logger.info("Initializing Marker OCR engine")
        # In a real implementation, we would initialize the OCR library here
    
    async def extract_text(self, file_path: Path) -> OCRResult:
        """Extract text from an image using OCR.
        
        Args:
            file_path: Path to the image file
            
        Returns:
            OCR result
        """
        start_time = time.time()
        warnings = []
        
        try:
            # Check if file exists
            if not file_path.exists():
                warnings.append(f"File not found: {file_path}")
                return OCRResult(
                    text="",
                    confidence=0.0,
                    engine_used=OCREngine.MARKER,
                    processing_time=time.time() - start_time,
                    warnings=warnings
                )
            
            # Check if file format is supported
            if not self.supports_format(file_path):
                warnings.append(f"Unsupported file format: {file_path.suffix}")
                return OCRResult(
                    text="",
                    confidence=0.0,
                    engine_used=OCREngine.MARKER,
                    processing_time=time.time() - start_time,
                    warnings=warnings
                )
            
            # Open image
            image = Image.open(file_path)
            
            # In a real implementation, we would use an OCR library here
            # For now, we'll return a placeholder result
            
            # This is just a placeholder - in a real implementation,
            # we would extract text from the image using an OCR library
            text = f"[Placeholder OCR text for {file_path.name}]"
            confidence = 0.7  # Placeholder confidence score
            
            # Create result
            result = OCRResult(
                text=text,
                confidence=confidence,
                engine_used=OCREngine.MARKER,
                processing_time=time.time() - start_time,
                warnings=warnings
            )
            
            return result
        
        except Exception as e:
            logger.error(f"Error extracting text with Marker: {e}")
            warnings.append(f"Error extracting text: {str(e)}")
            
            return OCRResult(
                text="",
                confidence=0.0,
                engine_used=OCREngine.MARKER,
                processing_time=time.time() - start_time,
                warnings=warnings
            )
    
    def get_name(self) -> str:
        """Get the name of the OCR engine.
        
        Returns:
            Name of the OCR engine
        """
        return "Marker"
    
    def get_version(self) -> str:
        """Get the version of the OCR engine.
        
        Returns:
            Version of the OCR engine
        """
        return "0.1.0"  # Placeholder version
    
    def get_supported_formats(self) -> Set[str]:
        """Get the supported file formats.
        
        Returns:
            Set of supported file extensions
        """
        return {
            ".jpg", ".jpeg", ".png", ".bmp", ".tiff", ".tif", ".gif"
        }
